Podział na Window i Scene
