+++ Riku/Riku/Riku.vcxproj
@@ -172,6 +172,7 @@
     <ClCompile Include="..\src\Frontend\Object.cpp" />
     <ClCompile Include="..\src\Frontend\Scene\Light.cpp" />
     <ClCompile Include="..\src\Frontend\Scene\Scene.cpp" />
+    <ClCompile Include="..\src\Frontend\Scene\Window.cpp" />
     <ClCompile Include="..\src\Frontend\Shader.cpp" />
     <ClCompile Include="..\src\Frontend\stb_image.cpp" />
     <ClCompile Include="..\src\Frontend\Texture.cpp" />
@@ -241,6 +242,7 @@
     <ClInclude Include="..\src\Frontend\Scene\Callbacks\ScrollCallback.h" />
     <ClInclude Include="..\src\Frontend\Scene\Light.h" />
     <ClInclude Include="..\src\Frontend\Scene\Scene.h" />
+    <ClInclude Include="..\src\Frontend\Scene\Window.h" />
     <ClInclude Include="..\src\Frontend\Shader.h" />
     <ClInclude Include="..\src\Frontend\stb_image.h" />
     <ClInclude Include="..\src\Frontend\Texture.h" />
