+++ Riku/src/Frontend/Frontend.cpp
@@ -1,10 +1,10 @@
 #include "Frontend.h"
 
 front::Frontend::Frontend(GameLogic& logic)
-	: logic(logic), state(logic), scene(config,logic,state)
+	: logic(logic), state(logic), window(config,logic,state)
 {}
 
 bool front::Frontend::update()
 {
-	return scene.update();
+	return window.update();
 }
