+++ Riku/src/Frontend/Window/Scene.h
@@ -1,35 +1,32 @@
 #pragma once
 #include <GL/glew.h>
 #include <GLFW/glfw3.h>
-#include "../Config.h"
 #include <vector>
 #include <glm/gtx/rotate_vector.hpp>
-#include "../Object.h"
-#include "../Shader.h"
-#include "../GUIFactory.h"
 #ifdef _WIN32
 #undef max
 #undef min
 #endif
+#include "../Object.h"
+#include "../Shader.h"
+#include "../GUIFactory.h"
+#include "../Config.h"
 #include "../Model.h"
 #include "Light.h"
 namespace front {
+	class Window;
 	class Scene
 	{
-		GameLogic& logic;
 		FrontendState& state;
-		GLFWwindow* window;
 		Config& config;
 
 		Transform movingCameraTransform;
-		float aspect;
 
 		Shader lightingShader;
 
-		float lastFrame = 0.0f;
-		float deltaTime;
 		float fogDensity = 0.01f;
 		int focusedUnitIndex = 0;
+		float& aspect;
 
 		Light light;
 		CEGUI::GUIFactory fac;
@@ -41,16 +38,13 @@ namespace front {
 		std::map<std::string, Model> objectModels;
 		Model unitModel;
 
-		void initWindow();
-		void setCallbacks();
-		std::optional<glm::vec2> getRelativeCursorPosition();
-		void processInput();
+		void init(GLFWwindow* window);
 	public:
-		Scene(Config& config, GameLogic& logic, FrontendState& state);
-		~Scene();
-		bool update();//input
+		Scene(Config& config, GameLogic& logic, FrontendState& state, float& aspect);
+		void update();//input
 		void draw();
 
+		friend class Window;
 	};
 }
